rsh-ai-backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
pinecone-client>=2.2.1
openai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
gunicorn>=21.0.0
tiktoken>=0.4.0
